Consolidate dialect directories
