+++ src/enzyme_ad/jax/Dialect/Comm/CommInterfaces.cpp
@@ -1,0 +1,6 @@
+#include "src/enzyme_ad/jax/Dialect/Comm/CommDialect.h"
+
+using namespace mlir;
+using namespace mlir::comm;
+
+#include "src/enzyme_ad/jax/Dialect/Comm/CommInterfaces.cpp.inc"
