--- src/enzyme_ad/jax/Dialects/Comm/CommInterfaces.cpp
@@ -1,6 +1,0 @@
-#include "src/enzyme_ad/jax/Dialects/Comm/CommDialect.h"
-
-using namespace mlir;
-using namespace mlir::comm;
-
-#include "src/enzyme_ad/jax/Dialects/Comm/CommInterfaces.cpp.inc"
